Did put X-Platform compatible code into utils.go
